debug

+++ CS3343_Project_Pilot/src/cs2115Handler/NORHandler.java
@@ -6,7 +6,7 @@ public class NORHandler extends LogicHandler{
 
 	@Override
 	public double handleForDoubleReturn(String expression) throws InvalidExpressionException {
-		return new NOTHandler().handleForDoubleReturn(""+new ORHandler().handleForDoubleReturn(expression));
+		return new NOTHandler().handleForDoubleReturn(""+(int)new ORHandler().handleForDoubleReturn(expression));
 	}
 
 }
